requirements.txt
aerich==0.8.2
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiojobs==1.3.0
aiomysql==0.2.0
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
async-timeout==5.0.1
asyncclick==8.1.8.0
attrs==25.1.0
bcrypt==4.3.0
captcha==0.7.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dictdiffer==0.9.0
ecdsa==0.19.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.11
frozenlist==1.5.0
h11==0.14.0
idna==3.10
iso8601==2.1.0
multidict==6.1.0
passlib==1.7.4
pillow==11.1.0
propcache==0.3.0
pyasn1==0.4.8
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyMySQL==1.1.1
pypika-tortoise==0.5.0
python-dotenv==1.0.1
python-jose==3.4.0
pytz==2025.1
redis==5.2.1
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tortoise-orm==0.24.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
